requirements.txt
boto3==1.26.90
PyGithub==1.58.1
jinja2==3.1.2
pytest==7.3.0
